requirements.txt
python-dotenv
PyPDF2
pdf2image
pytesseract
langchain
langchain-community
faiss-cpu
openai